test(Lotto) : 로또 1개 발행 시 오름차순 정렬 확인용 테스트 추가
